+++ encapsia_cli/config.py
@@ -25,6 +25,7 @@ def save(obj, output):
     api = lib.get_api(**obj)
     config = api.get_all_config()
     lib.pretty_print(config, "json", output=output)
+    lib.log_output(f"Configuration saved to {output.name}")
 
 
 @main.command()
@@ -35,6 +36,7 @@ def load(obj, input):
     api = lib.get_api(**obj)
     data = lib.parse(input.read(), "json")
     api.set_config_multi(data)
+    lib.log_output(f"Configuration loaded from {input.name}")
 
 
 @main.command()
@@ -61,6 +63,8 @@ def set(obj, key, value):
     api = lib.get_api(**obj)
     value = lib.parse(value, "json")
     api.set_config(key, value)
+    new_value = api.get_config(key)
+    lib.log_output(f"Configuration entry {key} has been set to {new_value}")
 
 
 @main.command()
@@ -70,3 +74,4 @@ def delete(obj, key):
     """Delete value against given key."""
     api = lib.get_api(**obj)
     api.delete_config(key)
+    lib.log_output(f"Configuration entry {key} deleted")