+++ encapsia_cli/plugininfo.py
@@ -318,8 +318,8 @@ class PluginInfos:
             if (p := PluginInfos(pis).latest()) is not None
         )
 
-    def filter_out_prereleases(self, include_dev_builds=False) -> PluginInfos:
-        if not include_dev_builds:
+    def filter_out_prereleases(self, include_prereleases=False) -> PluginInfos:
+        if not include_prereleases:
             return PluginInfos(pi for pi in self.pis if not pi.semver.prerelease)
         return self