+++ encapsia_cli/database.py
@@ -30,7 +30,7 @@ def backup(obj, filename):
 @main.command()
 @click.argument("filename", type=click.Path(exists=True))
 @click.option(
-    "--force", is_flag=True, help="[Deprecated] Don't prompt the user for confirmation."
+    "--force", is_flag=True, help="[DEPRECATED] Don't prompt the user for confirmation."
 )
 @click.option("--yes", is_flag=True, help="Don't prompt the user for confirmation.")
 @click.pass_obj