+++ walkthrough_tests/database.sh
@@ -10,7 +10,7 @@ encapsia database backup /tmp/test_encapsia_backup
 ls -l /tmp/test_encapsia_backup
 
 test "Restore from the backup"
-encapsia database restore --force /tmp/test_encapsia_backup
+encapsia database restore --yes /tmp/test_encapsia_backup
 sleep 10
 # Cleanup
 rm /tmp/test_encapsia_backup