+++ walkthrough_tests/plugins.sh
@@ -9,7 +9,7 @@ cd $(dirname "$0")
 start_tests
 
 test "Build the walkthrough_test plugin from src"
-encapsia plugins --force dev-build test_plugin
+encapsia plugins dev-build --overwrite test_plugin
 
 test "Request a build again, but this time it should be skipped over because it already exists in the local store"
 encapsia plugins dev-build test_plugin
@@ -18,12 +18,15 @@ test "Move the walkthrough_test plugin out of the local store and then add it ba
 mv ~/.encapsia/plugins/plugin-walkthrough_test-0.0.1.tar.gz /tmp/
 encapsia plugins add file:///tmp/plugin-walkthrough_test-0.0.1.tar.gz
 
+test "Install Manager (ensure we have a clean way to call the refresh-roles task)"
+encapsia plugins install --yes manager
+
 test "Install the walkthrough_test plugin from the local store, then uninstall it"
-encapsia plugins --force install --versions=example.toml --show-logs
-encapsia plugins --force uninstall walkthrough_test --show-logs
+encapsia plugins install --yes --versions=example.toml --show-logs
+encapsia plugins uninstall --yes walkthrough_test --show-logs
 
 test "Dev update the walkthrough_test plugin from scratch"
-encapsia plugins --force dev-update test_plugin
+encapsia plugins dev-update --all test_plugin
 
 test "Second time round there is nothing to do because nothing has changed"
 encapsia plugins dev-update test_plugin
@@ -34,8 +37,8 @@ encapsia plugins dev-update test_plugin
 rm test_plugin/tasks/test_new_module.py
 
 test "Install the non-dev version of the walkthrough_test plugin so we can uninstall it (to be tidy)."
-encapsia plugins --force install --versions=example.toml
-encapsia plugins --force uninstall walkthrough_test
+encapsia plugins install --yes --versions=example.toml
+encapsia plugins uninstall --yes walkthrough_test
 
 test "Create and destroy new namespace"
 encapsia plugins dev-create testing123
@@ -43,16 +46,16 @@ encapsia plugins dev-destroy testing123
 
 test "Fetch a few specific plugins from S3, install launch, and show the logs"
 encapsia plugins add --versions example_plugins.toml
-encapsia plugins --force install launch
+encapsia plugins install --yes launch
 encapsia plugins logs launch
 
 test "Fetch latest plugins from S3, install launch, and show the logs"
 encapsia plugins add --latest-existing
-encapsia plugins --force install launch
+encapsia plugins install --yes launch
 encapsia plugins logs launch
 
 test "Uninstall launch and show the logs"
-encapsia plugins --force uninstall launch
+encapsia plugins uninstall --yes launch
 encapsia plugins logs launch
 
 test "Get status on all installed plugins"
@@ -67,5 +70,5 @@ encapsia plugins upstream
 encapsia plugins upstream conduct-1.6 --all-versions
 
 test "Install a plugin using a path"
-encapsia plugins --force install example_plugins/plugin-foo-1.0.0.tar.gz
+encapsia plugins install --yes example_plugins/plugin-foo-1.0.0.tar.gz